+++ src/TrailDBEvent.java
@@ -3,7 +3,7 @@ package traildb;
 import traildb.TrailDBItem;
 
 public class TrailDBEvent {
-	public int timestamp;
-	public int numItems;
+	public long timestamp;
+	public long numItems;
 	TrailDBItem items[];
 }
